filtros cambiados, quedo 10 puntos
