+++ client/src/components/Home/home.module.css
@@ -14,10 +14,10 @@
 .errors{
     display: flex;
     background-color: #1C1C1C;
-    height: 100vh;
+    height: 100%;
+    width: 100%;
     flex-direction: column;
     justify-content: center;
-    width: 80%;
 }
 
 .button{
